refactor!: enforce mandatory variables by removing checks
By removing checks, we enforce the usage of some variables that are
deemed to be mandatory.

This commit also remove section data checks. So if user don't include
the respective section data, it implies the user don't want to include
the data and they need to exclude the respective section function call
in the template in order to compile.